fix: CI failures (prettier drift, Pyodide int32 overflow, singular-matrix test)
- Reformat 24 pre-existing Python READMEs to current prettier markdown
  emphasis style (*x* -> _x_); these were unrelated to this PR's content
  but blocked the merged CI check. Rebuild generated-curriculum.json to
  pick up the reformatted text.
- Fix vectorized_sum_of_squares/benchmark_sum_of_squares to cast to
  int64 before squaring: np.arange defaults to int32 on 32-bit/WASM
  platforms (Pyodide), so squaring values near 1,000,000 silently
  overflowed and produced a wrong sum in the browser test runner.
- Make the singular-matrix test for np.linalg.solve tolerant of
  backends (e.g. Pyodide's LAPACK build) that don't raise LinAlgError
  for an exactly-singular matrix, instead asserting the returned
  'solution' is non-finite or fails to satisfy the system.

+++ TrenTorch_Web/data/app_data/01-numpy/09-performance-memory/04-measuring-vectorized-vs-loop-performance/solution.py
@@ -11,7 +11,8 @@ def loop_sum_of_squares(values: list) -> int:
 
 
 def vectorized_sum_of_squares(arr: np.ndarray) -> int:
-    return (arr * arr).sum()
+    arr = arr.astype(np.int64)
+    return int((arr * arr).sum())
 
 
 def best_time(func, args: tuple, repeats: int) -> float:
@@ -25,7 +26,7 @@ def best_time(func, args: tuple, repeats: int) -> float:
 
 def benchmark_sum_of_squares(n: int, repeats: int = 3) -> dict:
     values = list(range(n))
-    arr = np.arange(n)
+    arr = np.arange(n, dtype=np.int64)
 
     loop_time = best_time(loop_sum_of_squares, (values,), repeats)
     vec_time = best_time(vectorized_sum_of_squares, (arr,), repeats)
